[sns] Expose only basic SNS ops
